app/requirements.txt
fastapi
uvicorn
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
python-multipart
